requirements.txt
Django==1.10.2
django-dotenv==1.4.1
djangorestframework==3.4.7
psycopg2==2.6.2
pySMART==0.3
SQLAlchemy==1.1.1
